+++ ARAnalytics.h
@@ -1,31 +1,3 @@
-//
-//  ARAnalytics.h
-//  Art.sy
-//
-//  Created by Orta Therox on 18/12/2012.
-//  Copyright (c) 2012 - Present Orta Therox & Art.sy. All rights reserved.
-//
-//  Permission is hereby granted, free of charge, to any person
-//  obtaining a copy of this software and associated documentation
-//  files (the "Software"), to deal in the Software without
-//  restriction, including without limitation the rights to use,
-//  copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the
-//  Software is furnished to do so, subject to the following
-//  conditions:
-//
-//  The above copyright notice and this permission notice shall be
-//  included in all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-//  OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-//  HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-//  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-//  FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
-//  OTHER DEALINGS IN THE SOFTWARE.
-
 // For OS X support we need to mock up UIVIewController/UINavigationViewController
 
 #if !TARGET_OS_IPHONE
@@ -63,6 +35,16 @@
 
 @interface ARAnalytics : NSObject <UINavigationControllerDelegate>
 
+/**
+ *  A flag that setup ARAnalytics to print it's log (ARLog) to stdout or not.
+ *
+ *  By default, ARLog collects log messages to providers and print them on the stdout.
+ *  If you want make the stdout clean while developing, you could set this flag to NO.
+ *
+ *  @param shouldPrint If YES, the ARLog will also print log message to stdout
+ */
++ (void)logShouldPrintStdout:(BOOL)shouldPrint;
+
 /// A global setup analytics API, keys are provided at the bottom of the documentation.
 + (void)setupWithAnalytics:(NSDictionary *)analyticsDictionary;
 
@@ -90,6 +72,12 @@
 + (void)setupChartbeatWithApplicationID:(NSString *)appID;
 + (void)setupLibratoWithEmail:(NSString *)email token:(NSString *)token prefix:(NSString *)prefix;
 + (void)setupSegmentioWithWriteKey:(NSString*)key;
++ (void)setupSwrveWithAppID:(NSString *)appID apiKey:(NSString *)apiKey;
++ (void)setupYandexMobileMetricaWithAPIKey:(NSString*)key;
++ (void)setupAdjustWithAppToken:(NSString *)token;
++ (void)setupBranchWithAPIKey:(NSString *)key;
++ (void)setupSnowplowWithAddress:(NSString *)address;
+
 
 /// Add a provider manually
 + (void)setupProvider:(ARAnalyticalProvider *)provider;
@@ -100,10 +88,10 @@
 /// Show all current providers
 + (NSSet *)currentProviders;
 
-
-/// Set a per user property
-/// @warning Deprecated, will be removed in next major release
-+ (void)identifyUserwithID:(NSString *)userID andEmailAddress:(NSString *)email __attribute__((deprecated));
+/// Get the instance of provider class which is setup ready.
+/// Developer must setup this provider ready via above methods and the argument must be a subclass of
+/// ARAnalyticalProvider or this methid returns nil.
++ (ARAnalyticalProvider *)providerInstanceOfClass:(Class)ProviderClass;
 
 /// Register a user and an associated email address, it is fine to send nils for either.
 + (void)identifyUserWithID:(NSString *)userID andEmailAddress:(NSString *)email;
@@ -119,6 +107,10 @@
 
 /// Submit user events to providers with additional properties
 + (void)event:(NSString *)event withProperties:(NSDictionary *)properties;
+
+/// Adds super properties, these are properties that are sent along with
+/// in addition to the event properties.
++ (void)addEventSuperProperties:(NSDictionary *)superProperties;
 
 /// Submit errors to providers
 + (void)error:(NSError *)error;
@@ -150,7 +142,7 @@
 @end
 
 /// an NSLog-like command that send to providers
-extern void ARLog (NSString *format, ...);
+extern void ARLog (NSString *format, ...) NS_FORMAT_FUNCTION(1,2);
 
 /// A try-catch for nil protection wrapped event
 extern void ARAnalyticsEvent (NSString *event, NSDictionary *properties);
@@ -187,3 +179,11 @@ extern const NSString *ARLibratoEmail;
 extern const NSString *ARLibratoToken;
 extern const NSString *ARLibratoPrefix;
 extern const NSString *ARSegmentioWriteKey;
+extern const NSString *ARSwrveAppID;
+extern const NSString *ARSwrveAPIKey;
+extern const NSString *ARYandexMobileMetricaAPIKey;
+extern const NSString *ARAdjustAppTokenKey;
+extern const NSString *ARAppsFlyerAppID;
+extern const NSString *ARAppsFlyerDevKey;
+extern const NSString *ARBranchAPIKey;
+extern const NSString *ARSnowplowURL;